fix : rename common utils to utils in remote package
