+++ app/src/main/java/com/cairosquad/movio/di/remoteDataSourceModule.kt
@@ -1,9 +1,9 @@
 package com.cairosquad.movio.di
 
-import com.cairosquad.remote.common.HttpClientFactory
-import com.cairosquad.remote.common.HttpEngine
 import com.cairosquad.remote.search.RemoteMovieDiscoveryDataSourceImpl
 import com.cairosquad.remote.search.RemoteSearchDataSourceImpl
+import com.cairosquad.remote.utils.HttpClientFactory
+import com.cairosquad.remote.utils.HttpEngine
 import com.cairosquad.repository.search.data_source.remote.RemoteMovieDiscoveryDataSource
 import com.cairosquad.repository.search.data_source.remote.RemoteSearchDataSource
 import org.koin.dsl.module
