+++ remote/src/main/java/com/cairosquad/remote/search/RemoteMovieDiscoveryDataSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.cairosquad.remote.search
 
 import com.cairosquad.remote.BuildConfig
-import com.cairosquad.remote.common.utils.callApi
-import com.cairosquad.remote.common.utils.constructUrl
+import com.cairosquad.remote.utils.callApi
+import com.cairosquad.remote.utils.constructUrl
 import com.cairosquad.repository.search.data_source.remote.RemoteMovieDiscoveryDataSource
 import com.cairosquad.repository.search.data_source.remote.dto.MovieRemoteDto
 import com.cairosquad.repository.search.data_source.remote.dto.SearchResultResponse
