+++ remote/src/main/java/com/cairosquad/remote/utils/HttpClientFactory.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.remote.common
+package com.cairosquad.remote.utils
 
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
